refactor(lsm+persist): move codec module from persist to lsm, flip dep direction
Task 2: Move CrcProof, CodecRegistry, CodecError into minkowski-lsm::codec.
CrcProof and CodecRegistry::decode visibility widen to pub since they now
cross the crate boundary. ComponentSchema also moves to lsm::codec (from
persist::record) so that CodecRegistry::build_remap has no upward dep on
persist. persist::record re-exports ComponentSchema from lsm for continuity.

Task 3: Add minkowski-lsm as a dependency of minkowski-persist. Rewrite all
use crate::codec::* as use minkowski_lsm::codec::*. Re-export from
persist::lib for downstream convenience. No behavioral changes;
pure import rewrite.

Note: plan called for two commits with an intentionally broken intermediate
state, but workspace-level pre-commit hooks prevent committing broken state.
Combined into one clean commit.
