+++ crates/minkowski-lsm/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod codec;
 pub mod error;
 pub mod format;
 pub mod manifest;
